♻️ refactor(device-store): keep store→feature dependency one-directional
The device store action was importing feature-layer modules (deviceCwd +
recentDirs), reversing the store←feature flow.

- Move `deviceCwd` (WorkingDirEntry + nextWorkingDirs/removeWorkingDir, pure
  device-domain logic) into `src/store/device`; features now import it from
  `@/store/device`.
- Keep localStorage recents (`recentDirs`) in the feature layer: the migration
  hook moves there, reads/clears localStorage, and passes the entries *into* the
  store action — the store no longer touches feature storage.
