+++ src/routes/(main)/settings/devices/features/DeviceDetailPanel.tsx
@@ -8,10 +8,10 @@ import { FolderOpenIcon, FolderPlusIcon, XIcon } from 'lucide-react';
 import { memo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { nextWorkingDirs } from '@/features/ChatInput/RuntimeConfig/deviceCwd';
 import { renderDirIcon } from '@/features/ChatInput/RuntimeConfig/dirIcon';
 import { lambdaQuery } from '@/libs/trpc/client';
 import { electronSystemService } from '@/services/electron/system';
+import { nextWorkingDirs } from '@/store/device';
 
 import type { DeviceListItem } from './DeviceItem';
 import { getDeviceIcon } from './getDeviceIcon';
